move WriteBlockAndSetHeadWithTime to blockchain_arbitrum
